requirements.txt
flask
python-dotenv
supabase
qrcode
pydantic
psycopg2-binary
requests
gunicorn
